requirements.txt
fastapi
pydantic
typing
uvicorn
python-multipart
pandas
numpy
scipy
scikit-learn
openai
python-dotenv
xgboost